Use doubles for tempo, sample rate
Technically, most sample rates will cause an overflow in the float type.
Also, there's no reason not to use double precision internally.

+++ source/SampleSource.h
@@ -62,7 +62,7 @@ typedef struct {
   SampleSourceOpenAs openedAs;
   CharString sourceName;
   int numChannels;
-  float sampleRate;
+  double sampleRate;
   unsigned long numFramesProcessed;
 
   OpenSampleSourceFunc openSampleSource;
